[CHORE] : edit user_point type

+++ app/src/main/java/com/dosopt/naverpay/domain/model/benefit/User.kt
@@ -2,5 +2,5 @@ package com.dosopt.naverpay.domain.model.benefit
 
 data class User(
     val user_name : String = "",
-    val user_point : Int = 0,
+    val user_point : Long = 0,
 )
